refactors mobile/desktop navs
